admin_service/requirements.txt
bcrypt
mysql-connector-python
python-dotenv
pandas
langchain
langchain-mongodb
langchain-openai
datetime
google-cloud-storage
PyMuPDF
python-docx
openpyxl
pdfplumber
python-magic
flask
flask-login
requests
numpy==1.24.4